# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for trainer account, client, and workout management.",
        "title": "TrainDesk API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/reports/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Report summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period start, RFC3339",
                        "name": "started_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Period end, RFC3339",
                        "name": "ended_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.ReportSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DTO.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "DTO.ReportClientActivity": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "total_duration_min": {
                    "type": "integer"
                },
                "workout_count": {
                    "type": "integer"
                }
            }
        },
        "DTO.ReportSummaryResponse": {
            "type": "object",
            "properties": {
                "active_clients": {
                    "type": "integer"
                },
                "by_client": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.ReportClientActivity"
                    }
                },
                "by_type": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.ReportWorkoutTypeSummary"
                    }
                },
                "ended_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "total_duration_min": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "number"
                },
                "total_workouts": {
                    "type": "integer"
                }
            }
        },
        "DTO.ReportWorkoutTypeSummary": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}